[REFACTOR] Move Info component to shared component
